Create install_OPatch.sh
This script will automatically update the OPatch directory for the local Oracle installation. It backs up the current OPatch directory to OPatch_old and copies the response file (assuming the file name is the default ocm.rsp) from the old directory to the new one.

To do: take response file as a parameter, do an "if -e responsefile" check to prevent a file-not-found error, for running in a new environment or when a response file hasn't been created, or just use a flag to copy/not copy it.